feat: add basic markup and styles
