+++ package.json
@@ -12,11 +12,8 @@
     "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
     "lint:fix": "eslint . --ext js,ts,tsx --fix",
     "lint:staged": "lint-staged",
-    "test": "npx vitest run --coverage",
-    "test:run": "vitest run",
-    "test:coverage": "npx vitest --coverage",
     "prepare": "husky install",
-    "pp": "npm run format:fix && npm run lint:fix && npm run test"
+    "pp": "npm run format:fix && npm run lint:fix"
   },
   "lint-staged": {
     "src/**/*.{js,ts,tsx,scss,css,json}": "prettier --write",
@@ -33,16 +30,11 @@
     "@commitlint/cli": "^19.3.0",
     "@commitlint/config-conventional": "^19.2.2",
     "@eslint/js": "^9.6.0",
-    "@testing-library/jest-dom": "^6.4.6",
-    "@testing-library/react": "^16.0.0",
-    "@testing-library/user-event": "^14.5.2",
     "@types/react": "^18.3.3",
     "@types/react-dom": "^18.3.0",
     "@typescript-eslint/eslint-plugin": "^7.15.0",
     "@typescript-eslint/parser": "^7.15.0",
     "@vitejs/plugin-react": "^4.3.1",
-    "@vitest/coverage-v8": "^1.6.0",
-    "@vitest/ui": "^1.6.0",
     "eslint": "^8.57.0",
     "eslint-config-airbnb": "^19.0.4",
     "eslint-config-airbnb-typescript": "^18.0.0",
@@ -56,13 +48,10 @@
     "eslint-plugin-react-refresh": "^0.4.7",
     "globals": "^15.8.0",
     "husky": "^9.0.11",
-    "jsdom": "^24.1.0",
     "lint-staged": "^15.2.7",
-    "msw": "^2.3.2",
     "prettier": "^3.3.2",
     "typescript": "^5.2.2",
     "typescript-eslint": "^7.15.0",
-    "vite": "^5.3.1",
-    "vitest": "^1.6.0"
+    "vite": "^5.3.1"
   }
 }
