+++ src/components/header.tsx
@@ -1,18 +1,23 @@
 import { NavLink } from "react-router-dom";
+import "./header.css";
 
 function Header() {
   return (
-    <div className="title" data-testid="title">
-      <h1>React Forms</h1>
-      <NavLink to="/regular" className="link">
-        Regular Forms
-      </NavLink>
-      <NavLink to="/controlled" className="link">
-        Controlled Forms
-      </NavLink>
-      <NavLink to="/" className="link">
-        Home
-      </NavLink>
+    <div className="header">
+      <div className="title">
+        <h1>React Forms</h1>
+      </div>
+      <nav className="navigation">
+        <NavLink to="/regular" className="link">
+          Regular Forms
+        </NavLink>
+        <NavLink to="/controlled" className="link">
+          Controlled Forms
+        </NavLink>
+        <NavLink to="/" className="link">
+          Home
+        </NavLink>
+      </nav>
     </div>
   );
 }
