+++ src/main.css
@@ -20,6 +20,7 @@
   --box-shadow-color: #d0d0d0;
   --link-color: #808080;
   --link-disabled-color: #b1b1b1;
+  --link-active-color: #e7e7e7;
   --flyout-color: #213547;
   --flyout-background-color: #bdbdbd;
 }
@@ -28,34 +29,9 @@
   width: 100vw;
   min-height: 100vh;
   display: flex;
-}
-
-.root-theme {
-  width: 100vw;
-  min-height: 100vh;
-  display: flex;
-  text-align: center;
   flex-direction: column;
+  justify-content: flex-start;
   align-items: center;
-  padding: 0 2rem;
-  max-height: 100vh;
-}
-
-.root-theme[data-theme="Dark"] {
-  --color: #cfcfcf;
-  --background-color: #313131;
-  --a-color: #d4d4d4;
-  --a-hover-color: #e6e6e6;
-  --button-background-color: #1d1d1d;
-  --button-hover-border-color: #a3a3a3;
-  --error-boundary-color: #ff8a8a;
-  --error-boundary-body-color: #999999;
-  --border-color: #535353;
-  --box-shadow-color: #494949;
-  --link-color: #979797;
-  --link-disabled-color: #646464;
-  --flyout-color: #213547;
-  --flyout-background-color: #979797;
 }
 
 * {
@@ -87,25 +63,4 @@ h1 {
   font-size: 2.2em;
   line-height: 1.1;
   margin: 10px auto;
-}
-
-button {
-  border-radius: 4px;
-  border: 1px solid transparent;
-  padding: 0.4em 1.2em;
-  font-size: 1em;
-  font-weight: 500;
-  font-family: inherit;
-  background-color: var(--button-background-color);
-  cursor: pointer;
-  transition: border-color 0.25s;
-}
-
-button:hover {
-  border-color: var(--button-hover-border-color);
-}
-
-button:focus,
-button:focus-visible {
-  outline: 4px auto -webkit-focus-ring-color;
 }
