--- src/page.css
@@ -1,261 +1,0 @@
-#root {
-  display: flex;
-  text-align: center;
-  flex-direction: column;
-  align-items: center;
-  margin: 0 auto;
-  padding: 0 2rem;
-}
-
-.error-page {
-  color: var(--error-boundary-color);
-  background-color: var(--error-boundary-body-color);
-}
-
-.error-boundary {
-  color: var(--error-boundary-color);
-  max-width: 780px;
-  border: 1px solid var(--error-boundary-color);
-  border-radius: 8px;
-  padding: 4px;
-  margin: 10% auto;
-}
-
-.error-boundary-head {
-  color: var(--error-boundary-color);
-  font-weight: bold;
-  font-size: 1.5rem;
-}
-
-.error-boundary-body {
-  color: var(--error-boundary-body-color);
-  max-width: 360px;
-  text-wrap: wrap;
-  overflow-wrap: anywhere;
-  margin: 20px auto;
-  padding: 20px;
-  background-color: var(--error-boundary-color);
-}
-
-.title {
-  display: flex;
-  position: relative;
-}
-
-.theme-change {
-  position: absolute;
-  right: 8px;
-  top: 8px;
-  border-color: var(--box-shadow-color);
-  border-radius: 4px;
-  border: 1px solid transparent;
-  padding: 0.4em 1.2em;
-  font-size: 1.3em;
-  font-weight: 600;
-  font-family: inherit;
-  background-color: var(--button-background-color);
-  cursor: pointer;
-  transition: border-color 0.25s;
-}
-
-.theme-change:hover {
-  transform: scale(1.03);
-  border-color: var(--button-hover-border-color);
-}
-
-.item-checked-change {
-  position: absolute;
-  right: 0px;
-  top: 0px;
-  font-size: larger;
-  box-shadow: 2px 2px 9px 0px var(--box-shadow-color);
-  border-radius: 4px;
-  padding: 0 4px;
-  cursor: pointer;
-}
-
-.character-checked-change {
-  position: absolute;
-  left: 0px;
-  top: 0px;
-  font-size: larger;
-  box-shadow: 2px 2px 9px 0px var(--box-shadow-color);
-  border-radius: 4px;
-  padding: 0 4px;
-  cursor: pointer;
-  font-size: inherit;
-}
-
-.change-input {
-  cursor: pointer;
-}
-
-.spinner {
-  height: 200px;
-  width: 200px;
-  padding: 1.5em;
-  margin-top: 40px;
-  background-color: transparent;
-}
-
-.spinner-img {
-  height: 100%;
-}
-
-.search-query {
-  width: 100%;
-  min-width: 320px;
-  max-width: 640px;
-}
-
-.search-query-input {
-  width: 70%;
-  border-radius: 4px;
-  border: 1px solid var(--border-color);
-  padding: 0.4em 0.2em;
-  margin: 0 0.4em;
-  font-size: 1em;
-  font-weight: 500;
-  font-family: inherit;
-}
-
-.search-result {
-  position: relative;
-  display: flex;
-  flex-direction: row;
-  justify-content: flex-start;
-  margin-top: 10px;
-  padding: 20px;
-  min-width: 320px;
-  max-width: 1440px;
-  gap: 20px;
-  box-shadow: 2px 2px 5px 3px var(--box-shadow-color);
-  align-items: stretch;
-  max-height: 80vh;
-  overflow-y: clip;
-  flex-wrap: nowrap;
-  height: 80vh;
-}
-
-.search-result-empty {
-  display: flex;
-  flex-direction: column;
-  margin-top: 40px;
-  padding: 20px;
-  min-width: 320px;
-  box-shadow: 2px 2px 5px 3px var(--box-shadow-color);
-}
-
-.search-result-list {
-  display: flex;
-  flex-direction: row;
-  flex-wrap: wrap;
-  justify-content: center;
-  align-content: flex-start;
-  margin-top: 10px;
-  padding: 10px 10px;
-  width: 100%;
-  min-width: 320px;
-  max-width: 1000px;
-  max-height: 100%;
-  gap: 20px;
-  overflow-y: scroll;
-}
-
-.item {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: flex-start;
-  width: 320px;
-  box-shadow: 2px 2px 5px 3px var(--box-shadow-color);
-  border-radius: 4px;
-  padding: 2px 6px;
-}
-
-.item-button {
-  position: relative;
-}
-
-.item-button:hover {
-  transform: scale(1.01);
-  box-shadow: 2px 2px 5px 3px var(--color);
-}
-
-.item-name {
-  font-weight: bold;
-}
-
-.card {
-  display: flex;
-  flex-direction: row;
-  align-content: flex-start;
-  align-items: flex-start;
-  justify-content: flex-start;
-  height: fit-content;
-  height: 68vh;
-  font-size: x-large;
-  padding: 20px 20px;
-  box-shadow: 2px 2px 5px 3px var(--box-shadow-color);
-  border-radius: 4px;
-  gap: 10px;
-}
-
-.card-left {
-  overflow: hidden;
-  height: 100%;
-}
-
-.card-left-img {
-  height: 100%;
-  max-height: 30vw;
-}
-
-.card-right {
-  position: relative;
-  width: 320px;
-  display: flex;
-  flex-direction: column;
-  gap: 4vh;
-}
-
-.card-spinner {
-  margin-top: 20%;
-  margin-left: 1%;
-  background-color: transparent;
-}
-
-.card-line {
-  width: 100%;
-  display: flex;
-}
-
-.card-name {
-  font-weight: bold;
-}
-
-.card-button-close {
-  align-self: end;
-  font-size: 1rem;
-}
-
-.link {
-  color: var(--link-color);
-  padding: 4px 8px;
-  box-shadow: 2px 2px 5px 3px var(--box-shadow-color);
-  border-radius: 4px;
-  border: 1px solid transparent;
-}
-
-.link:hover {
-  border-radius: 4px;
-  border: 1px solid var(--border-color);
-}
-
-.disabled-link {
-  color: var(--link-disabled-color);
-  pointer-events: none;
-  cursor: default;
-  border: 1px solid transparent;
-  box-shadow: none;
-}
