+++ src/pages/notFound.css
@@ -1,0 +1,5 @@
+.not-found {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
