+++ src/pages/submits.tsx
@@ -1,4 +1,4 @@
-import Page from "../page";
+import Page from "../wrapper";
 
 function Submits() {
   return (
