+++ src/wrapper.css
@@ -1,0 +1,40 @@
+.root-element {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  flex-wrap: nowrap;
+  justify-content: space-between;
+  width: 100vw;
+  height: 100vh;
+}
+
+.title {
+  display: flex;
+  position: relative;
+}
+
+.link {
+  color: var(--link-color);
+  padding: 4px 8px;
+  box-shadow: 2px 2px 5px 3px var(--box-shadow-color);
+  border-radius: 4px;
+  border: 1px solid transparent;
+}
+
+.link:hover {
+  color: var(--link-color);
+  border: 1px solid var(--border-color);
+}
+
+.link.active {
+  border: 1px solid var(--border-color);
+  background-color: var(--link-active-color);
+}
+
+.disabled-link {
+  color: var(--link-disabled-color);
+  pointer-events: none;
+  cursor: default;
+  border: 1px solid transparent;
+  box-shadow: none;
+}
